Responded to feedback on PR
